+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Redefined "Display Brightness" as "Hardware Brightness", and "Image Brightness" as "Software Brightness".
 - Updated overlay sidecar to .NET 8.0
+- Updated sleep detection to include controller button presses as a factor
 
 ### Fixed